pallet_moonbeam_foreign_assets

Module pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

The main components of this pallet are:

Pallet, which implements all of the dispatchable extrinsics of the pallet, among other public functions.

The subset of the functions that are dispatchable can be identified either in the dispatchables module or in the Call enum.

storage_types, which contains the list of all types that are representing a storage item. Otherwise, all storage items are listed among Type Definitions.

Config, which contains the configuration trait of this pallet.

Event and Error, which are listed among the Enums.

Modules

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Structs

AssetDepositDetails

GenesisConfig

Can be used to configure the genesis state of this pallet.

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Error

An error that can occur while executing the mapping pallet’s logic.

Event

The Event enum of this pallet

Constants

PALLET_ID

The moonbeam foreign assets’s pallet id

Traits

Config

Configuration trait of this pallet.

Type Aliases

AssetBalance

AssetId

AssetsById

Mapping from an asset id to a Foreign asset type. This is mostly used when receiving transaction specifying an asset directly, like transferring an asset from this chain to another.

AssetsByLocation

Reverse mapping of AssetsById. Mapping from a foreign asset to an asset id. This is mostly used when receiving a multilocation XCM message to retrieve the corresponding asset in which tokens should me minted.

AssetsCreationDetails

Mapping from an asset id to its creation details

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.