Trait pallet_moonbeam_foreign_assets::pallet::Config

pub trait Config: Config + Config + TypeInfo {

    type AccountIdToH160: Convert<Self::AccountId, H160>;
    type AssetIdFilter: Contains<AssetId>;
    type EvmRunner: Runner<Self>;
    type ConvertLocation: ConvertLocation<Self::AccountId>;
    type ForeignAssetCreatorOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = OriginType>;
    type ForeignAssetFreezerOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = OriginType>;
    type ForeignAssetModifierOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = OriginType>;
    type ForeignAssetUnfreezerOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = OriginType>;
    type OnForeignAssetCreated: ForeignAssetCreatedHook<Location>;
    type MaxForeignAssets: Get<u32>;
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type WeightInfo: WeightInfo;
    type XcmLocationToH160: ConvertLocation<H160>;
    type ForeignAssetCreationDeposit: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
    type Balance: Member + Parameter + AtLeast32BitUnsigned + Default + Copy + MaybeSerializeDeserialize + MaxEncodedLen + TypeInfo;
    type Currency: ReservableCurrency<Self::AccountId>;

}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type AccountIdToH160: Convert<Self::AccountId, H160>

type AssetIdFilter: Contains<AssetId>

A filter to forbid some AssetId values, if you don’t use it, put “Everything”

type EvmRunner: Runner<Self>

EVM runner

type ConvertLocation: ConvertLocation<Self::AccountId>

type ForeignAssetCreatorOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = OriginType>

Origin that is allowed to create a new foreign assets

type ForeignAssetFreezerOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = OriginType>

Origin that is allowed to freeze all tokens of a foreign asset

type ForeignAssetModifierOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = OriginType>

Origin that is allowed to modify asset information for foreign assets

type ForeignAssetUnfreezerOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = OriginType>

Origin that is allowed to unfreeze all tokens of a foreign asset that was previously frozen

type OnForeignAssetCreated: ForeignAssetCreatedHook<Location>

Hook to be called when new foreign asset is registered.

type MaxForeignAssets: Get<u32>

Maximum numbers of different foreign assets

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

The overarching event type.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type XcmLocationToH160: ConvertLocation<H160>

type ForeignAssetCreationDeposit: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

Amount of tokens required to lock for creating a new foreign asset

type Balance: Member + Parameter + AtLeast32BitUnsigned + Default + Copy + MaybeSerializeDeserialize + MaxEncodedLen + TypeInfo

The balance type for locking funds

type Currency: ReservableCurrency<Self::AccountId>

The currency type for locking funds

Object Safety

This trait is not object safe.